+++ flytekit-jackson/src/main/java/org/flyte/flytekit/jackson/SdkTypeModule.java
@@ -20,6 +20,11 @@ import com.fasterxml.jackson.core.Version;
 import com.fasterxml.jackson.databind.Module;
 import com.fasterxml.jackson.databind.deser.Deserializers;
 import com.fasterxml.jackson.databind.module.SimpleSerializers;
+import org.flyte.flytekit.jackson.deserializers.LiteralMapDeserializers;
+import org.flyte.flytekit.jackson.deserializers.SdkBindingDataDeserializers;
+import org.flyte.flytekit.jackson.serializers.BindingMapSerializers;
+import org.flyte.flytekit.jackson.serializers.LiteralMapSerializers;
+import org.flyte.flytekit.jackson.serializers.SdkBindingDataSerializer;
 
 class SdkTypeModule extends Module {
   private static final Deserializers DEFAULT_SDKBINDING_DESERIALIZERS =
