+++ flytekit-jackson/src/main/java/org/flyte/flytekit/jackson/deserializers/CustomSdkBindingDataDeserializer.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2021 Flyte Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.flyte.flytekit.jackson.deserializers;
+
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.TreeNode;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+import com.fasterxml.jackson.databind.node.TextNode;
+import java.io.IOException;
+import java.util.Map;
+import org.flyte.flytekit.SdkBindingData;
+
+class CustomSdkBindingDataDeserializer extends JsonDeserializer<SdkBindingData<?>> {
+  private final Map<String, SdkBindingData<?>> bindingsMap;
+
+  public CustomSdkBindingDataDeserializer(Map<String, SdkBindingData<?>> bindingsMap) {
+    this.bindingsMap = bindingsMap;
+  }
+
+  @Override
+  public SdkBindingData<?> deserialize(
+      JsonParser jsonParser, DeserializationContext deserializationContext) throws IOException {
+    TreeNode treeNode = jsonParser.readValueAsTree();
+    String attr = ((TextNode) treeNode).asText();
+    return bindingsMap.get(attr);
+  }
+}
