+++ flytekit-jackson/src/main/java/org/flyte/flytekit/jackson/deserializers/LiteralMapDeserializers.java
@@ -14,7 +14,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.flyte.flytekit.jackson;
+package org.flyte.flytekit.jackson.deserializers;
 
 import com.fasterxml.jackson.databind.BeanDescription;
 import com.fasterxml.jackson.databind.DeserializationConfig;
@@ -22,21 +22,18 @@ import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.deser.Deserializers;
 import java.util.Map;
-import org.flyte.flytekit.SdkBindingData;
+import org.flyte.api.v1.LiteralType;
+import org.flyte.flytekit.jackson.JacksonLiteralMap;
 
-// TODO find a better name
-class CustomSdkBindingDataDeserializers extends Deserializers.Base {
-  private final Map<String, SdkBindingData<?>> bindingsMap;
-
-  CustomSdkBindingDataDeserializers(Map<String, SdkBindingData<?>> bindingsMap) {
-    this.bindingsMap = bindingsMap;
-  }
+public class LiteralMapDeserializers extends Deserializers.Base {
 
   @Override
   public JsonDeserializer<?> findBeanDeserializer(
       JavaType type, DeserializationConfig config, BeanDescription beanDesc) {
-    if (SdkBindingData.class.isAssignableFrom(type.getRawClass())) {
-      return new CustomSdkBindingDataDeserializer(bindingsMap);
+    if (type.getRawClass().equals(JacksonLiteralMap.class)) {
+      Map<String, LiteralType> literalTypeMap = type.getValueHandler();
+
+      return new LiteralMapDeserializer(literalTypeMap);
     }
 
     return null;
