+++ flytekit-jackson/src/main/java/org/flyte/flytekit/jackson/serializers/FloatSerializer.java
@@ -14,32 +14,33 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.flyte.flytekit.jackson;
+package org.flyte.flytekit.jackson.serializers;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import java.io.IOException;
-import org.flyte.api.v1.Blob;
 import org.flyte.api.v1.Literal;
 import org.flyte.api.v1.LiteralType;
-import org.flyte.api.v1.Scalar;
+import org.flyte.api.v1.Primitive;
+import org.flyte.api.v1.SimpleType;
 
-public class BlobSerializer extends ScalarSerializer {
-  public BlobSerializer(
+public class FloatSerializer extends PrimitiveSerializer {
+
+  public FloatSerializer(
       JsonGenerator gen,
       String key,
       Literal value,
       SerializerProvider serializerProvider,
       LiteralType literalType) {
     super(gen, key, value, serializerProvider, literalType);
+    if (literalType.getKind() != LiteralType.Kind.SIMPLE_TYPE
+        && literalType.simpleType() != SimpleType.FLOAT) {
+      throw new IllegalArgumentException("Literal type should be a float literal type");
+    }
   }
 
   @Override
-  void serializeScalar() throws IOException {
-    gen.writeFieldName("scalar");
-    gen.writeObject(Scalar.Kind.BLOB);
-    serializerProvider
-        .findValueSerializer(Blob.class)
-        .serialize(value.scalar().blob(), gen, serializerProvider);
+  public void serializePrimitive() throws IOException {
+    writePrimitive(Primitive.Kind.FLOAT_VALUE, (gen, value) -> gen.writeNumber(value.floatValue()));
   }
 }
