+++ flytekit-jackson/src/main/java/org/flyte/flytekit/jackson/serializers/LiteralSerializer.java
@@ -14,33 +14,41 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.flyte.flytekit.jackson;
+package org.flyte.flytekit.jackson.serializers;
+
+import static org.flyte.flytekit.jackson.util.JacksonConstants.LITERAL;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import java.io.IOException;
 import org.flyte.api.v1.Literal;
 import org.flyte.api.v1.LiteralType;
-import org.flyte.api.v1.Primitive;
-import org.flyte.api.v1.SimpleType;
 
-public class FloatSerializer extends PrimitiveSerializer {
+abstract class LiteralSerializer {
 
-  public FloatSerializer(
+  protected final JsonGenerator gen;
+  protected final String key;
+  protected final Literal value;
+  protected final SerializerProvider serializerProvider;
+  protected final LiteralType literalType;
+
+  public LiteralSerializer(
       JsonGenerator gen,
       String key,
       Literal value,
       SerializerProvider serializerProvider,
       LiteralType literalType) {
-    super(gen, key, value, serializerProvider, literalType);
-    if (literalType.getKind() != LiteralType.Kind.SIMPLE_TYPE
-        && literalType.simpleType() != SimpleType.FLOAT) {
-      throw new IllegalArgumentException("Literal type should be a float literal type");
-    }
+    this.gen = gen;
+    this.key = key;
+    this.value = value;
+    this.serializerProvider = serializerProvider;
+    this.literalType = literalType;
   }
 
-  @Override
-  public void serializePrimitive() throws IOException {
-    writePrimitive(Primitive.Kind.FLOAT_VALUE, (gen, value) -> gen.writeNumber(value.floatValue()));
+  final void serialize() throws IOException {
+    gen.writeFieldName(LITERAL);
+    serializeLiteral();
   }
+
+  abstract void serializeLiteral() throws IOException;
 }
